+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+* Updated dataset.
 * Minor code improvements.
 
 v0.68.3 (2025-03-26)